switching to symbolic types
