+++ client/valor/schemas/__init__.py
@@ -1,44 +1,8 @@
-from .constraints import (
-    BoolMapper,
-    Constraint,
-    DictionaryMapper,
-    GeometryMapper,
-    GeospatialMapper,
-    LabelMapper,
-    NumericMapper,
-    StringMapper,
-)
 from .evaluation import EvaluationParameters, EvaluationRequest
 from .filters import Filter
-from .geometry import (
-    BasicPolygon,
-    BoundingBox,
-    MultiPolygon,
-    Point,
-    Polygon,
-    Raster,
-)
-from .metadata import dump_metadata, load_metadata, validate_metadata
 
 __all__ = [
-    "BasicPolygon",
-    "Point",
-    "BoundingBox",
-    "Polygon",
-    "MultiPolygon",
-    "Raster",
     "EvaluationRequest",
     "EvaluationParameters",
     "Filter",
-    "validate_metadata",
-    "dump_metadata",
-    "load_metadata",
-    "Constraint",
-    "BoolMapper",
-    "NumericMapper",
-    "StringMapper",
-    "GeometryMapper",
-    "GeospatialMapper",
-    "DictionaryMapper",
-    "LabelMapper",
 ]
